chore: Move deprecated code from src/ to deprecated/ and move and rename wcdbasemodel to waribasemodel
