+++ deprecated/english_reference.py
@@ -1,0 +1,5 @@
+# from src.models.wikimedia.wikipedia.reference.generic import WikipediaReference
+
+
+# class EnglishWikipediaReference(WikipediaReference):
+#     pass
