+++ deprecated/wcd_item.py
