--- src/models/api/schema/__init__.py
@@ -1,5 +1,0 @@
-from marshmallow import Schema, fields
-
-
-class RefreshSchema(Schema):
-    refresh = fields.Bool(required=False)
