+++ src/models/api/schema/article_schema.py
@@ -3,7 +3,7 @@ import logging
 from marshmallow import fields, post_load
 
 from src.models.api.job.article_job import ArticleJob
-from src.models.api.schema import RefreshSchema
+from src.models.api.schema.refresh import RefreshSchema
 
 logger = logging.getLogger(__name__)
 
@@ -18,7 +18,7 @@ class ArticleSchema(RefreshSchema):
     # **kwargs is needed here despite what the validator claims
     def return_object(self, data, **kwargs) -> ArticleJob:  # type: ignore # dead: disable
         """Return job object"""
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("return_object: running")
         job = ArticleJob(**data)
