+++ src/models/api/schema/check_doi_schema.py
@@ -2,7 +2,7 @@ from marshmallow import post_load
 from marshmallow.fields import Int, String
 
 from src.models.api.job.check_doi_job import CheckDoiJob
-from src.models.api.schema import RefreshSchema
+from src.models.api.schema.refresh import RefreshSchema
 
 
 class CheckDoiSchema(RefreshSchema):
@@ -16,7 +16,7 @@ class CheckDoiSchema(RefreshSchema):
     # **kwargs is needed here despite what the validator claims
     def return_object(self, data, **kwargs) -> CheckDoiJob:  # type: ignore # dead: disable
         """Return job object"""
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("return_object: running")
         job = CheckDoiJob(**data)
