+++ src/models/api/schema/check_url_schema.py
@@ -4,7 +4,7 @@ from marshmallow import post_load
 from marshmallow.fields import Int, String
 
 from src.models.api.job.check_url_job import CheckUrlJob
-from src.models.api.schema import RefreshSchema
+from src.models.api.schema.refresh import RefreshSchema
 
 logger = logging.getLogger(__name__)
 
@@ -20,7 +20,7 @@ class CheckUrlSchema(RefreshSchema):
     # **kwargs is needed here despite what the validator claims
     def return_object(self, data, **kwargs) -> CheckUrlJob:  # type: ignore # dead: disable
         """Return job object"""
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("return_object: running")
         job = CheckUrlJob(**data)
