+++ src/models/api/schema/refresh.py
@@ -1,0 +1,5 @@
+from marshmallow import Schema, fields
+
+
+class RefreshSchema(Schema):
+    refresh = fields.Bool(required=False)
