+++ src/models/base/__init__.py
@@ -7,14 +7,14 @@ from pydantic import BaseModel, validate_arguments
 logger = logging.getLogger(__name__)
 
 
-class WcdBaseModel(BaseModel):
+class WariBaseModel(BaseModel):
     """This base model has all methods that we
     want to use in more than one class"""
 
     # We set to Any here because of cyclic dependency or pydantic forward ref error
     cache: Optional[Any] = None
 
-    class Config:
+    class Config:  # dead: disable
         extra = "forbid"  # dead: disable
 
     @validate_arguments
