+++ src/models/base/job.py
@@ -1,11 +1,11 @@
 from typing import Optional
 
 from src.models.api.job import Job
-from src.wcd_base_model import WcdBaseModel
+from src.models.base import WariBaseModel
 
 
-class JobBaseModel(WcdBaseModel):
+class JobBaseModel(WariBaseModel):
     job: Optional[Job] = None
 
-    class Config:
+    class Config:  # dead: disable
         extra = "forbid"  # dead: disable
