+++ src/models/wikimedia/wikipedia/analyzer.py
@@ -6,14 +6,14 @@ from typing import Any, Dict, List, Optional
 from src.models.api.job.article_job import ArticleJob
 from src.models.api.statistic.article import ArticleStatistics
 from src.models.api.statistic.reference import ReferenceStatistic
+from src.models.base import WariBaseModel
 from src.models.exceptions import MissingInformationError
 from src.models.wikimedia.wikipedia.article import WikipediaArticle
-from src.wcd_base_model import WcdBaseModel
 
 logger = logging.getLogger(__name__)
 
 
-class WikipediaAnalyzer(WcdBaseModel):
+class WikipediaAnalyzer(WariBaseModel):
     """This model contain all the logic for getting the
     article and reference statistics and mapping them to the API output model
 
@@ -102,7 +102,7 @@ class WikipediaAnalyzer(WcdBaseModel):
 
     def __analyze__(self):
         """Helper method"""
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("__analyze__: running")
         if self.job:
@@ -112,7 +112,7 @@ class WikipediaAnalyzer(WcdBaseModel):
                 self.article.fetch_and_extract_and_parse()
 
     def __gather_reference_statistics__(self):
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("__gather_reference_statistics__: running")
         if (
@@ -157,7 +157,7 @@ class WikipediaAnalyzer(WcdBaseModel):
             )
 
     def __populate_article__(self):
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("__populate_article__: running")
         if self.job and self.job.title:
