--- src/models/wikimedia/wikipedia/reference/english/english_reference.py
@@ -1,5 +1,0 @@
-from src.models.wikimedia.wikipedia.reference.generic import WikipediaReference
-
-
-class EnglishWikipediaReference(WikipediaReference):
-    pass
