+++ src/models/wikimedia/wikipedia/reference/generic.py
@@ -6,7 +6,7 @@ from typing import Any, Dict, List, Optional, Union
 from mwparserfromhell.nodes import Tag  # type: ignore
 from mwparserfromhell.wikicode import Wikicode  # type: ignore
 
-from src.models.basemodels.job import JobBaseModel
+from src.models.base.job import JobBaseModel
 from src.models.exceptions import MissingInformationError
 from src.models.wikimedia.wikipedia.reference.enums import (
     FootnoteSubtype,
@@ -169,7 +169,7 @@ class WikipediaReference(JobBaseModel):
 
     def __extract_first_level_domains__(self) -> None:
         """This aggregates all first level domains from the urls found in the raw references"""
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("__extract_first_level_domains__: running")
         if not self.reference_urls_done:
@@ -230,7 +230,7 @@ class WikipediaReference(JobBaseModel):
 
     def __extract_templates_and_parameters__(self) -> None:
         """Helper method"""
-        from src.models.api import app
+        from src import app
 
         app.logger.debug(
             "__extract_templates_and_parameters_from_raw_reference__: running"
@@ -241,7 +241,7 @@ class WikipediaReference(JobBaseModel):
 
     def __extract_raw_templates__(self) -> None:
         """Extract the templates from self.wikicode"""
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("__extract_raw_templates__: running")
         if not self.wikicode:
@@ -281,7 +281,7 @@ class WikipediaReference(JobBaseModel):
 
     def __extract_and_clean_template_parameters__(self) -> None:
         """We extract all templates"""
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("__extract_and_clean_template_parameters__: running")
         [
@@ -291,7 +291,7 @@ class WikipediaReference(JobBaseModel):
 
     def extract_and_check(self) -> None:
         """Helper method"""
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("extract_and_check: running")
         self.__extract_templates_and_parameters__()
