+++ src/views/check_doi.py
@@ -30,7 +30,7 @@ class CheckDoi(StatisticsWriteView):
     def get(self):
         """This is the main method and the entrypoint for flask
         Every branch in this method has to return a tuple (Any,response_code)"""
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("get: running")
         self.__validate_and_get_job__()
@@ -38,7 +38,7 @@ class CheckDoi(StatisticsWriteView):
             return self.__handle_valid_job__()
 
     def __handle_valid_job__(self):
-        from src.models.api import app
+        from src import app
 
         app.logger.debug("__handle_valid_job__; running")
