+++ wsgi.py
@@ -1,3 +1,3 @@
 """This is the entry point for gunicorn"""
 # noinspection PyUnresolvedReferences
-from src.models.api import app
+from src import app
